organize interfaces / types (#6)
* organize interfaces / types


* organize interfaces / types


---------
