--- src/components/player/types/index.tsx
@@ -1,1 +1,0 @@
-export * from './Track'
