+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './Artist'
+export * from './Player'
+export * from './Track'
